+++ src/main/java/in/koreatech/koin/domain/owner/repository/OwnerAttachmentRepository.java
@@ -4,9 +4,9 @@ import org.springframework.data.repository.Repository;
 
 import in.koreatech.koin.domain.owner.model.OwnerAttachment;
 
-public interface OwnerAttachmentRepository extends Repository<OwnerAttachment, Long> {
+public interface OwnerAttachmentRepository extends Repository<OwnerAttachment, Integer> {
 
     OwnerAttachment save(OwnerAttachment ownerAttachment);
 
-    OwnerAttachment findById(Long id);
+    OwnerAttachment findById(Integer id);
 }
