+++ src/main/java/in/koreatech/koin/domain/owner/repository/OwnerRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.Repository;
 import in.koreatech.koin.domain.owner.exception.OwnerNotFoundException;
 import in.koreatech.koin.domain.owner.model.Owner;
 
-public interface OwnerRepository extends Repository<Owner, Long> {
+public interface OwnerRepository extends Repository<Owner, Integer> {
 
     Optional<Owner> findById(Integer ownerId);
