+++ src/main/java/in/koreatech/koin/domain/user/repository/UserRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.Repository;
 import in.koreatech.koin.domain.user.exception.UserNotFoundException;
 import in.koreatech.koin.domain.user.model.User;
 
-public interface UserRepository extends Repository<User, Long> {
+public interface UserRepository extends Repository<User, Integer> {
 
     User save(User user);
